split metrics into vis_metrics and datacontainer
