+++ hera_qm/__init__.py
@@ -2,4 +2,5 @@
 #import vis_metrics
 #import cal_metrics
 import xrfi
-import metrics
+import vis_metrics
+import datacontainer
